Add TTS interface and Azure model with metrics
Created subfolder /Robot/ for storing all systems related strictly to the robot. Created a TTS interface that can be used to swap our TTS model if latency is an issue. Implemented Azure TTS model with hard coded Azure key for now for testing, will need to create a config file later in order to allow deployability and not charge my Azure account s**t ton of money. Created metrics that can be used for dashboards in order to monitor TTS latency and create error logs if needed.