Add LLM module with Azure OpenAI and OpenAI engines
Added LLM module for Azure OpenAI and OpenAI. Azure is set to default as it is being used for STT and TTS and offers free student credits. OpenAI base engine can be swapped in if latency or cost gets too high on Azure. Kept same layout as TTS and STT with interface allowing for new engines to be built out. Implemented conversation context manager for remembering past conversations and allowing LLM to contain a state.