Add Conversation Orchestrator to connect all modules
Added conversation_orchestrator.py to tie all modules together and run a conversational mode on an html browser. Added recording timer that is acting as a place holder, informing the user how long they have to speak to the robot. Added a test that runs through some basic sounds/phrases, demonstrating the phonetic capabilities of the animated mouth. Made final adjustments in preparation of the demo and prototype assignment coming up.